+++ docs/whats_new.md
@@ -1,5 +1,9 @@
 # What's New
 
+## Unreleased
+
+* Fix kerchunk JSON file filtering.
+
 ## v0.12.2 (April 15, 2025)
 
 * fixed issue when setting `vertical_mixing` False for `OpenOil` was not passed through correctly